Move ISchemaSourceSelector to the Core project
