+++ SyncKusto.Core/Models/ValidationResult.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace SyncKusto.Core.Models;
+
+/// <summary>
+/// Result of a validation operation
+/// </summary>
+public record ValidationResult(
+    bool IsValid,
+    string? ErrorMessage = null)
+{
+    public static ValidationResult Success() => new(true);
+    public static ValidationResult Failure(string errorMessage) => new(false, errorMessage);
+}
