+++ SyncKusto/Models/SchemaSourceInfo.cs
@@ -3,30 +3,9 @@
 
 using SyncKusto.Core.Abstractions;
 using SyncKusto.Core.Models;
-using SyncKusto.SyncSources;
 using Kusto.Data.Common;
 
 namespace SyncKusto.Models;
-
-/// <summary>
-/// Information about a schema source (file or Kusto)
-/// </summary>
-public record SchemaSourceInfo(
-    SourceSelection SourceType,
-    string? FilePath = null,
-    KustoConnectionInfo? KustoInfo = null);
-
-/// <summary>
-/// Connection information for a Kusto database
-/// </summary>
-public record KustoConnectionInfo(
-    string Cluster,
-    string Database,
-    AuthenticationMode AuthMode,
-    string Authority,
-    string? AppId = null,
-    string? AppKey = null,
-    string? CertificateThumbprint = null);
 
 /// <summary>
 /// Result of a comparison operation
@@ -35,14 +14,3 @@ public record ComparisonResult(
     SchemaDifferenceResult Differences,
     DatabaseSchema SourceSchema,
     DatabaseSchema TargetSchema);
-
-/// <summary>
-/// Result of a validation operation
-/// </summary>
-public record ValidationResult(
-    bool IsValid,
-    string? ErrorMessage = null)
-{
-    public static ValidationResult Success() => new(true);
-    public static ValidationResult Failure(string errorMessage) => new(false, errorMessage);
-}
